https://datascientists.plan.io/issues/9813 - pushed quick changes , untested version of new JSON add book design

+++ src/main/java/org/occideas/book/service/BookService.java
@@ -1,6 +1,7 @@
 package org.occideas.book.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.occideas.book.dao.BookDao;
@@ -8,19 +9,27 @@ import org.occideas.book.dao.BookModuleDao;
 import org.occideas.book.mapper.BookMapper;
 import org.occideas.book.request.BookRequest;
 import org.occideas.book.response.BookVO;
+import org.occideas.config.BookConfig;
 import org.occideas.entity.Book;
 import org.occideas.entity.BookModule;
 import org.occideas.exceptions.BookNotExistException;
+import org.occideas.exceptions.GenericException;
 import org.occideas.node.dao.NodeDao;
 import org.occideas.security.handler.TokenManager;
 import org.occideas.security.model.TokenResponse;
+import org.occideas.utilities.FileUtil;
 import org.occideas.utilities.NodeUtil;
+import org.occideas.vo.NodeVO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.io.IOException;
+import java.nio.file.FileSystems;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -29,7 +38,7 @@ import java.util.Optional;
 @Service
 public class BookService {
 
-    private final Logger log = LogManager.getLogger(this.getClass());
+    private static Logger log = LogManager.getLogger(BookService.class);
 
     @Autowired
     private BookModuleDao bookModuleDao;
@@ -41,6 +50,8 @@ public class BookService {
     private NodeUtil nodeUtil;
     @Autowired
     private NodeDao nodeDao;
+    @Autowired
+    private BookConfig bookConfig;
 
 
     public List<BookVO> getBooks() {
@@ -68,62 +79,49 @@ public class BookService {
     }
 
     @Async
-    public void addModuleToBook(BookRequest bookRequest, String updatedBy) throws JsonProcessingException {
-
-        saveModuleToBook(bookRequest, updatedBy);
+    public void addModuleToBook(BookRequest bookRequest) throws JsonProcessingException {
         final Optional<Book> bookById = bookDao.findById(bookRequest.getBookId());
         if (bookById.isEmpty()) {
             throw new BookNotExistException("Book does not exist");
         }
         final Book book = bookById.get();
-        book.setLastUpdated(LocalDateTime.now());
-        bookDao.save(book);
+        String directory = bookConfig.getPath() + FileSystems.getDefault().getSeparator() + book.getName();
+        boolean success = FileUtil.saveJsonToFile(bookRequest.getJson(), directory, bookRequest.getName());
+        if (success) {
+            saveModuleToBook(bookRequest, directory + FileSystems.getDefault().getSeparator() + bookRequest.getName() + ".json");
+            book.setLastUpdated(LocalDateTime.now());
+            bookDao.save(book);
+        } else {
+            throw new GenericException("Unable to save to book. Error occured, check the logs.");
+        }
     }
 
-//    @Async
-//    public void addModuleToBookByNode(long idNode, String updatedBy) throws JsonProcessingException {
-//        NodeVO nodeVO = nodeUtil.convertToNodeVO(nodeDao.getNode(idNode));
-//        String json = new ObjectMapper().writeValueAsString(nodeVO);
-//
-//        final BookModule bookModule = new BookModule(
-//                bookRequest.getBookId(),
-//                nodeVO.getName(),
-//                valueAsBytes, nodeVO.hashCode(),
-//                nodeVO.getClass().getName(),
-//                updatedBy);
-//
-//        if (shouldSave) {
-//            log.info("Adding Module {}", nodeVO.getName());
-//            bookModuleDao.save(bookModule);
-//        }
-//
-//        nodeVO.getChildNodes().stream()
-//                .forEach(node -> {
-//                    try {
-//                        if (node.getLink() != 0l) {
-//                            saveModuleToBook(new BookRequest(node.getLink(), bookRequest.getBookId()), updatedBy, true);
-//                        } else {
-//                            saveModuleToBook(new BookRequest(node.getIdNode(), bookRequest.getBookId()), updatedBy, false);
-//                        }
-//                    } catch (JsonProcessingException e) {
-//                        log.error(e.getMessage(), e);
-//                    }
-//                });
-//    }
+    @Async
+    public void addModuleToBookByNode(long idNode, long bookId, String updatedBy) throws JsonProcessingException {
+        NodeVO nodeVO = nodeUtil.convertToNodeVO(nodeDao.getNode(idNode));
+        String json = new ObjectMapper().writeValueAsString(nodeVO);
+        addModuleToBook(new BookRequest(bookId, nodeVO.getName(), nodeVO.getType(), json, updatedBy));
+    }
 
-    private void saveModuleToBook(BookRequest bookRequest, String updatedBy) throws JsonProcessingException {
+    private void saveModuleToBook(BookRequest bookRequest, String jsonPath) throws JsonProcessingException {
         final BookModule bookModule = new BookModule(
                 bookRequest.getBookId(),
                 bookRequest.getName(),
-                bookRequest.getJson(),
+                jsonPath,
                 bookRequest.getType(),
-                updatedBy,
+                bookRequest.getUpdatedBy(),
                 LocalDateTime.now());
 
         bookModuleDao.save(bookModule);
     }
 
-    public void deleteBookModuleInBook(long bookId, String name) {
+    public void deleteBookModuleInBook(long bookId, String name) throws IOException {
+        Optional<BookModule> byNameAndBookId = bookModuleDao.findByNameAndBookId(name, bookId);
+        if (byNameAndBookId.isEmpty()) {
+            throw new GenericException("book module is no longer exist." + name);
+        }
+        log.info("Deleting module {}", byNameAndBookId.get().getJson());
+        Files.delete(Path.of(byNameAndBookId.get().getJson()));
         bookModuleDao.deleteByBookIdAndName(bookId, name);
     }
